Created tables and a superuser login.
Created all tables in mysite\Specialtopic\models.py, add all the tables to be viewed in by admin or superuser in mysite\Specialtopic\admin.py.

Also added a superuser, to access it run the manage.py file in shell using this command 'python manage.py runserver' and then add the '/admin' to the end of the link given in the shell.