ApScheduler library
For this to work you need to install ApScheduler -  $ pip install apscheduler

I have created and updater_excel which is scheduled at a particular time to run and also configured the app.py present in Specialtopic root directory.

This is import so the scheduler can run when the site is on.